Bug 95394 - Check update version in system and use the newer one if it exists. r=fabrice.desre
(cherry picked from commit 30f741c4fa3444070c7e07148f3df8bc24f093c7)

+++ updater-daemon.sh
@@ -1,15 +1,33 @@
 #!/system/bin/sh
 
-if [ ! -d /data/local/service/updater ]; then
+newver=`cat /system/kaios/updater.ver`
+oldver=`cat /data/local/service/updater/installed/updater-*.toml | grep version | cut -f 2 -d '"'`
+newer=0
+
+IFS="." read -a v_a <<< "$newver"
+IFS="." read -a v_b <<< "$oldver"
+
+i=0
+a=${v_a[$i]}
+b=${v_b[$i]}
+while [[ -n "$a" || -n "$b" ]]; do
+  [[ -z "$a" || "$a" -gt "$b" ]] && newer=1 && break
+
+  i=`expr $i + 1`
+  a=${v_a[$i]}
+  b=${v_b[$i]}
+done
+
+if [ ! -d /data/local/service/updater -o "$newer" = "1" ]; then
   mkdir -p /data/local/service/updater
   cp -f /system/kaios/updater-daemon /data/local/service/updater/
   cp -f /system/kaios/updater.toml /data/local/service/updater/
   chmod 755 /data/local/service/updater/updater-daemon
 fi
-if [ -d /system/kaios/test-fixtures ]; then
+if [ -d /system/kaios/test-fixtures -a "$newer" = "1" ]; then
   cp -rf /system/kaios/test-fixtures  /data/local/service/updater/
 fi
-if [ -f /system/kaios/updatercmd ]; then
+if [ -f /system/kaios/updatercmd -a "$newer" = "1" ]; then
   cp -f /system/kaios/updatercmd /data/local/service/updater/
   chmod 755 /data/local/service/updater/updatercmd
 fi
